refactor: audit §13 #5 — NativeVideoPresenter sub-controllers
Extract three @MainActor sub-controllers from NativeVideoPresenter.swift
(1,678 → 1,099 lines) into Shared/Screens/VideoPlayer/:

- PlaybackReporter — reportStart/Stop/Background + 10-tick throttled
  periodic progress. Fanned out from the presenter's shared time observer
  via onTick(). Context captured through a weak closure so itemId/player
  are always current.

- SkipSegmentController — intro/outro skip affordance. Platform-split UI
  (iOS floating UIButton / tvOS contextualActions). Now cancels in-flight
  fetches in teardown() so episode navigation can't race stale segments
  into the new episode.

- SleepTimerController — countdown indicator + "Still watching?" prompt
  (UIAlertController on tvOS, blur card on iOS). Takes a playerVCProvider
  closure and an onStopPlayback callback so the presenter stays the sole
  owner of AVPlayerViewController lifecycle.

Presenter retains the single addPeriodicTimeObserver (1 s) and fans out
ticks to skipSegments.onTick + playbackReporter.onTick, preserving the
one-observer invariant documented in CLAUDE.md. Finished-series overlay
and debug skip-to-end stay on the presenter (share sizing with the
remaining shared button styling).

Behavior-preserving: no visual / focus / localization / timing changes.
Deliberately keeps the pre-existing episode-nav reporting quirk (itemId
/ startTime are immutable) — flagged as a follow-up in Audit.md §17 so
the refactor commit stays reviewable.

iOS + tvOS builds green. 36/36 CinemaxTests pass.
