+++ Shared/Screens/NativeVideoPresenter.swift
@@ -40,6 +40,11 @@ final class NativeVideoPresenter {
     private var manifestLoader = HLSManifestLoader()
     private var backgroundObserver: NSObjectProtocol?
 
+    // Sub-controllers (see Shared/Screens/VideoPlayer/)
+    private var playbackReporter: PlaybackReporter!
+    private var skipSegments: SkipSegmentController!
+    private var sleepTimer: SleepTimerController!
+
     // Track state
     private var audioTracks: [MediaTrackInfo] = []
     private var subtitleTracks: [MediaTrackInfo] = []
@@ -47,33 +52,9 @@ final class NativeVideoPresenter {
     private var currentSubtitleIndex: Int? = nil
     private var currentPlayMethod: CinemaxKit.PlayMethod = .transcode
 
-    // Skip segments
-    //
-    // Pure time-based UX: the skip button is visible iff `currentTime ∈ segment`.
-    // No "already skipped" memory — if the user rewinds back into a segment, the
-    // button reappears.
-    //
-    // Platform split (see `showSkipButton` / `hideSkipButton` for the rendering):
-    // - iOS: a floating `UIButton` added directly to `AVPlayerViewController.view`
-    //   (stored in `skipButton`). Touch reaches it natively.
-    // - tvOS: the native `AVPlayerViewController.contextualActions` API. It's the
-    //   only mechanism that produces a focusable action button that coexists with
-    //   the transport-bar focus context. Custom subviews / overlay modals cannot
-    //   be focused by the Siri Remote while AVPlayerViewController is on screen —
-    //   the player's focus environment is private and locked.
-    private var segments: [MediaSegmentDto] = []
+    // Shared periodic time observer. Fans out to SkipSegmentController.onTick
+    // and PlaybackReporter.onTick from startProgressReporting.
     private var timeObserver: Any?
-    private var activeSegmentType: MediaSegmentType?
-    #if os(iOS)
-    private var skipButton: UIButton?
-    #endif
-
-    // Sleep timer
-    private var sleepTickTask: Task<Void, Never>?
-    private var sleepEndDate: Date?
-    private var sleepIndicatorContainer: UIView?
-    private var sleepIndicatorLabel: UILabel?
-    private var sleepOverlayContainer: UIView?
 
     // End-of-series
     private var currentSeriesName: String?
@@ -111,6 +92,23 @@ final class NativeVideoPresenter {
         self.autoPlayNextEpisode = autoPlayNextEpisode
         self.imageBuilder = imageBuilder
         self.onDismiss = onDismiss
+
+        self.playbackReporter = PlaybackReporter(
+            apiClient: apiClient, userId: userId,
+            context: { [weak self] in
+                guard let self, let info = self.playbackInfo else { return nil }
+                return .init(itemId: self.itemId, info: info, player: self.playerVC?.player)
+            }
+        )
+        self.skipSegments = SkipSegmentController(
+            apiClient: apiClient, loc: loc,
+            playerVCProvider: { [weak self] in self?.playerVC }
+        )
+        self.sleepTimer = SleepTimerController(
+            loc: loc,
+            playerVCProvider: { [weak self] in self?.playerVC },
+            onStopPlayback: { [weak self] in self?.playerVC?.dismiss(animated: true) }
+        )
     }
 
     func present(info: PlaybackInfo) {
@@ -156,7 +154,7 @@ final class NativeVideoPresenter {
         // Use AVPlayerViewControllerDelegate to detect dismissal (Menu button).
         let delegate = TVDismissDelegate()
         delegate.onDismiss = { [weak self] in
-            self?.reportPlaybackStop()
+            self?.playbackReporter.reportStop()
             self?.cleanup()
             self?.onDismiss()
         }
@@ -169,7 +167,7 @@ final class NativeVideoPresenter {
         let hostingVC = PlayerHostingVC(playerVC: vc)
         hostingVC.modalPresentationStyle = .fullScreen
         hostingVC.onDismissed = { [weak self] in
-            self?.reportPlaybackStop()
+            self?.playbackReporter.reportStop()
             self?.cleanup()
             self?.onDismiss()
         }
@@ -209,12 +207,12 @@ final class NativeVideoPresenter {
 
             avPlayer.replaceCurrentItem(with: playerItem)
             avPlayer.play()
-            reportPlaybackStart()
+            self.playbackReporter.reportStart(startTime: self.startTime)
             startProgressReporting()
             observeItemEnd(playerItem, player: avPlayer)
-            fetchSegments(for: self.itemId)
+            self.skipSegments.load(for: self.itemId)
             fetchAndApplyChapters(for: self.itemId, playerItem: playerItem)
-            startSleepTimerIfNeeded()
+            self.sleepTimer.startIfNeeded()
             showSkipToEndButtonIfDebugEnabled()
         }
     }
@@ -445,7 +443,7 @@ final class NativeVideoPresenter {
     private func navigateToEpisode(_ ep: EpisodeRef) {
         guard let navigator = episodeNavigator, let vc = playerVC else { return }
         Task {
-            reportPlaybackStop()
+            playbackReporter.reportStop()
             guard let (info, prev, next) = await navigator(ep.id) else { return }
             cleanupPlayer()
             self.hasRetriedDirectURL = false
@@ -477,13 +475,13 @@ final class NativeVideoPresenter {
             }
 
             avPlayer.play()
-            reportPlaybackStart()
+            playbackReporter.reportStart(startTime: self.startTime)
             startProgressReporting()
             observeItemEnd(playerItem, player: avPlayer)
-            fetchSegments(for: ep.id)
+            skipSegments.load(for: ep.id)
             fetchAndApplyChapters(for: ep.id, playerItem: playerItem)
             // Episode navigation restarts the sleep timer (keeps playback "session" alive).
-            startSleepTimerIfNeeded()
+            sleepTimer.startIfNeeded()
             showSkipToEndButtonIfDebugEnabled()
         }
     }
@@ -497,45 +495,14 @@ final class NativeVideoPresenter {
         item.externalMetadata = [meta]
     }
 
-    // MARK: - Playback Reporting
-
-    private func reportPlaybackStart() {
-        guard let info = playbackInfo else { return }
-        let positionTicks = startTime.map { Int($0 * 10_000_000) } ?? 0
-        let id = itemId
-        let client = apiClient
-        let uid = userId
-        Task.detached {
-            await client.reportPlaybackStart(
-                itemId: id, userId: uid,
-                mediaSourceId: info.mediaSourceId, playSessionId: info.playSessionId,
-                positionTicks: positionTicks, playMethod: info.playMethod
-            )
-        }
-    }
-
-    private func reportPlaybackStop() {
-        guard let info = playbackInfo else { return }
-        let positionTicks = Int((playerVC?.player?.currentTime().seconds ?? 0) * 10_000_000)
-        let id = itemId
-        let client = apiClient
-        let uid = userId
-        Task.detached {
-            await client.reportPlaybackStopped(
-                itemId: id, userId: uid,
-                mediaSourceId: info.mediaSourceId, playSessionId: info.playSessionId,
-                positionTicks: positionTicks
-            )
-        }
-    }
-
-    /// Single periodic time observer (0.5 s) that handles both segment skip detection
-    /// and playback progress reporting (every ~10 s).
-    private var progressTickCounter = 0
+    // MARK: - Shared Time Observer
+    //
+    // Single periodic observer (1 s) that fans out to segment skip detection and
+    // playback progress reporting. The reporter applies a 10-tick throttle.
 
     private func startProgressReporting() {
         removeTimeObserver()
-        progressTickCounter = 0
+        playbackReporter.resetTicking()
         guard let player = playerVC?.player else { return }
 
         timeObserver = player.addPeriodicTimeObserver(
@@ -544,32 +511,9 @@ final class NativeVideoPresenter {
         ) { [weak self] time in
             MainActor.assumeIsolated {
                 guard let self else { return }
-                // Segment skip check every 1 s
-                self.checkSegments(currentTime: time.seconds)
-
-                // Progress report every ~10 s (10 ticks × 1 s)
-                self.progressTickCounter += 1
-                if self.progressTickCounter >= 10 {
-                    self.progressTickCounter = 0
-                    self.reportPeriodicProgress()
-                }
+                self.skipSegments.onTick(currentTime: time.seconds)
+                self.playbackReporter.onTick()
             }
-        }
-    }
-
-    private func reportPeriodicProgress() {
-        guard let info = playbackInfo, let player = playerVC?.player else { return }
-        let ticks = Int(player.currentTime().seconds * 10_000_000)
-        let isPaused = player.rate == 0
-        let id = itemId
-        let client = apiClient
-        let uid = userId
-        Task.detached {
-            await client.reportPlaybackProgress(
-                itemId: id, userId: uid,
-                mediaSourceId: info.mediaSourceId, playSessionId: info.playSessionId,
-                positionTicks: ticks, isPaused: isPaused, playMethod: info.playMethod
-            )
         }
     }
 
@@ -594,161 +538,6 @@ final class NativeVideoPresenter {
         }
     }
 
-    // MARK: - Skip Segments
-
-    private func fetchSegments(for itemId: String) {
-        segments = []
-        hideSkipButton()
-        activeSegmentType = nil
-
-        Task { [weak self] in
-            guard let self else { return }
-            do {
-                let fetched = try await self.apiClient.getMediaSegments(
-                    itemId: itemId,
-                    includeSegmentTypes: [.intro, .outro]
-                )
-                self.segments = fetched
-            } catch {
-                logger.info("Media segments unavailable for \(itemId): \(error.localizedDescription)")
-            }
-        }
-    }
-
-    /// Drives skip-button visibility purely from `currentTime`:
-    /// - If we're inside an intro/outro segment that we're not already showing a
-    ///   button for, show it.
-    /// - If we're outside all segments but a button is shown, hide it.
-    ///
-    /// Re-entry is allowed: if the user rewinds back into a segment after leaving
-    /// it, `activeSegmentType` will have been cleared, so the next tick shows the
-    /// button again.
-    private func checkSegments(currentTime: Double) {
-        for segment in segments {
-            let start = Double(segment.startTicks ?? 0) / 10_000_000
-            let end = Double(segment.endTicks ?? 0) / 10_000_000
-            guard end > start else { continue }
-
-            if currentTime >= start && currentTime < end - 1 {
-                if activeSegmentType != segment.type {
-                    activeSegmentType = segment.type
-                    showSkipButton(for: segment)
-                }
-                return
-            }
-        }
-
-        // Outside all segments — clear the button if one is up.
-        if activeSegmentType != nil {
-            activeSegmentType = nil
-            hideSkipButton()
-        }
-    }
-
-    /// Renders the skip button. Platform-split:
-    /// - **iOS**: a floating `UIButton` added to `AVPlayerViewController.view`.
-    ///   Touch reaches it directly.
-    /// - **tvOS**: a `UIAction` installed on `AVPlayerViewController.contextualActions`.
-    ///   This is the native tvOS API for skip-intro-style affordances: the button
-    ///   appears in the player's own focus container, is focusable by the Siri
-    ///   Remote, and coexists with the transport bar (users can navigate from the
-    ///   scrubber/play button up to the skip action without losing HUD access).
-    ///   Custom subviews / overlay modals cannot achieve this — AVPlayerViewController
-    ///   locks its focus environment.
-    private func showSkipButton(for segment: MediaSegmentDto) {
-        guard let vc = playerVC else { return }
-
-        let title: String
-        switch segment.type {
-        case .intro:
-            title = loc.localized("player.skipIntro")
-        case .outro:
-            title = loc.localized("player.skipCredits")
-        default:
-            return
-        }
-
-        let endSeconds = Double(segment.endTicks ?? 0) / 10_000_000
-
-        #if os(tvOS)
-        // Native tvOS path. Setting `contextualActions` to a non-empty array makes
-        // the action visible in the playback chrome; clearing it removes the button.
-        // The time observer (via `checkSegments`) drives both sides, so the button's
-        // lifetime is exactly `[segment.start, segment.end)`.
-        let action = UIAction(
-            title: title,
-            image: UIImage(systemName: "forward.fill")
-        ) { [weak self] _ in
-            guard let player = self?.playerVC?.player else { return }
-            player.seek(
-                to: CMTime(seconds: endSeconds, preferredTimescale: 600),
-                toleranceBefore: .zero, toleranceAfter: .zero
-            )
-            // No manual hide call: the seek moves currentTime past segment.end,
-            // `checkSegments` sees we're outside all segments, and hideSkipButton
-            // clears the contextual action. Pure time-based state.
-        }
-        vc.contextualActions = [action]
-        #else
-        hideSkipButton() // idempotent
-
-        // Modern UIButton.Configuration replaces the deprecated `contentEdgeInsets`,
-        // `setTitle`, `setTitleColor`, etc. Visuals (frosted glass + 20% white tint)
-        // are reproduced via `background.customView` (UIVisualEffectView) plus
-        // `background.backgroundColor`.
-        var config = UIButton.Configuration.plain()
-        var attrTitle = AttributedString("  \(title)  ▶▶")
-        attrTitle.font = .systemFont(ofSize: buttonFontSize, weight: .semibold)
-        attrTitle.foregroundColor = UIColor.white
-        config.attributedTitle = attrTitle
-        config.contentInsets = NSDirectionalEdgeInsets(top: 0, leading: buttonPaddingH, bottom: 0, trailing: buttonPaddingH)
-
-        let blur = UIVisualEffectView(effect: UIBlurEffect(style: .dark))
-        blur.isUserInteractionEnabled = false
-        blur.layer.cornerRadius = buttonCornerRadius
-        blur.clipsToBounds = true
-        config.background.customView = blur
-        config.background.backgroundColor = UIColor.white.withAlphaComponent(0.2)
-        config.background.cornerRadius = buttonCornerRadius
-
-        let action = UIAction { [weak self] _ in
-            guard let player = self?.playerVC?.player else { return }
-            player.seek(
-                to: CMTime(seconds: endSeconds, preferredTimescale: 600),
-                toleranceBefore: .zero, toleranceAfter: .zero
-            )
-            // Same time-based cleanup as tvOS — the next checkSegments tick will
-            // hide the button when currentTime crosses segment.end.
-        }
-        let button = UIButton(configuration: config, primaryAction: action)
-        button.translatesAutoresizingMaskIntoConstraints = false
-        button.alpha = 0
-        self.skipButton = button
-
-        let targetView = vc.view!
-        targetView.addSubview(button)
-        NSLayoutConstraint.activate([
-            button.trailingAnchor.constraint(equalTo: targetView.safeAreaLayoutGuide.trailingAnchor, constant: -20),
-            button.bottomAnchor.constraint(equalTo: targetView.safeAreaLayoutGuide.bottomAnchor, constant: -80),
-            button.heightAnchor.constraint(equalToConstant: 44),
-        ])
-        UIView.animate(withDuration: 0.3) { button.alpha = 1 }
-        #endif
-    }
-
-    private func hideSkipButton() {
-        #if os(tvOS)
-        // Clearing `contextualActions` removes the button from the HUD.
-        playerVC?.contextualActions = []
-        #else
-        guard let button = skipButton else { return }
-        UIView.animate(withDuration: 0.25, animations: { button.alpha = 0 }) { _ in
-            button.removeFromSuperview()
-        }
-        skipButton = nil
-        #endif
-    }
-
     private func removeTimeObserver() {
         if let observer = timeObserver, let player = playerVC?.player {
             player.removeTimeObserver(observer)
@@ -756,6 +545,8 @@ final class NativeVideoPresenter {
         timeObserver = nil
     }
 
+    // Shared button styling for the debug Skip-to-End pill and the sleep
+    // indicator. Skip-intro/credits styling is owned by `SkipSegmentController`.
     private var buttonFontSize: CGFloat {
         #if os(tvOS)
         28
@@ -777,6 +568,48 @@ final class NativeVideoPresenter {
         32
         #else
         20
+        #endif
+    }
+
+    // Shared overlay styling (used by the iOS finishedSeriesOverlay).
+
+    private var overlayCardWidth: CGFloat {
+        #if os(tvOS)
+        640
+        #else
+        340
+        #endif
+    }
+
+    private var overlayIconSize: CGFloat {
+        #if os(tvOS)
+        72
+        #else
+        48
+        #endif
+    }
+
+    private var overlayTitleSize: CGFloat {
+        #if os(tvOS)
+        36
+        #else
+        22
+        #endif
+    }
+
+    private var overlaySubtitleSize: CGFloat {
+        #if os(tvOS)
+        22
+        #else
+        15
+        #endif
+    }
+
+    private var overlayButtonSize: CGFloat {
+        #if os(tvOS)
+        24
+        #else
+        16
         #endif
     }
 
@@ -1141,399 +974,6 @@ final class NativeVideoPresenter {
         #endif
     }
 
-    // MARK: - Sleep Timer
-
-    /// Reads the effective sleep-timer duration (user setting or debug override) and starts
-    /// a timer if non-zero. Called on playback start, episode navigation, and "Keep watching".
-    private func startSleepTimerIfNeeded() {
-        let seconds = SleepTimerOption.currentDefaultSeconds
-        guard seconds > 0 else { return }
-        startSleepTimer(seconds: seconds)
-    }
-
-    private func startSleepTimer(seconds: TimeInterval) {
-        stopSleepTimer()
-        sleepEndDate = Date().addingTimeInterval(seconds)
-        showSleepIndicator()
-        updateSleepIndicator(remaining: seconds)
-
-        sleepTickTask = Task { @MainActor [weak self] in
-            while !Task.isCancelled {
-                guard let self, let end = self.sleepEndDate else { return }
-                let remaining = end.timeIntervalSinceNow
-                if remaining <= 0 {
-                    self.stopSleepTimer()
-                    self.hideSleepIndicator()
-                    self.triggerSleep()
-                    return
-                }
-                self.updateSleepIndicator(remaining: remaining)
-                try? await Task.sleep(nanoseconds: 1_000_000_000)
-            }
-        }
-    }
-
-    private func stopSleepTimer() {
-        sleepTickTask?.cancel()
-        sleepTickTask = nil
-        sleepEndDate = nil
-    }
-
-    private func triggerSleep() {
-        playerVC?.player?.pause()
-        showSleepOverlay()
-    }
-
-    private func showSleepIndicator() {
-        guard sleepIndicatorContainer == nil, let vc = playerVC else { return }
-
-        let container = UIView()
-        container.translatesAutoresizingMaskIntoConstraints = false
-        container.layer.cornerRadius = buttonCornerRadius
-        container.clipsToBounds = true
-        container.alpha = 0
-
-        let blur = UIVisualEffectView(effect: UIBlurEffect(style: .dark))
-        blur.translatesAutoresizingMaskIntoConstraints = false
-        blur.isUserInteractionEnabled = false
-        container.addSubview(blur)
-
-        let icon = UIImageView(image: UIImage(systemName: "moon.zzz.fill"))
-        icon.tintColor = .white
-        icon.translatesAutoresizingMaskIntoConstraints = false
-        icon.contentMode = .scaleAspectFit
-
-        let label = UILabel()
-        label.translatesAutoresizingMaskIntoConstraints = false
-        label.font = .systemFont(ofSize: indicatorFontSize, weight: .semibold)
-        label.textColor = .white
-        label.text = ""
-
-        container.addSubview(icon)
-        container.addSubview(label)
-
-        let targetView = vc.view!
-        targetView.addSubview(container)
-
-        NSLayoutConstraint.activate([
-            blur.topAnchor.constraint(equalTo: container.topAnchor),
-            blur.bottomAnchor.constraint(equalTo: container.bottomAnchor),
-            blur.leadingAnchor.constraint(equalTo: container.leadingAnchor),
-            blur.trailingAnchor.constraint(equalTo: container.trailingAnchor),
-
-            icon.leadingAnchor.constraint(equalTo: container.leadingAnchor, constant: indicatorPaddingH),
-            icon.centerYAnchor.constraint(equalTo: container.centerYAnchor),
-            icon.widthAnchor.constraint(equalToConstant: indicatorIconSize),
-            icon.heightAnchor.constraint(equalToConstant: indicatorIconSize),
-
-            label.leadingAnchor.constraint(equalTo: icon.trailingAnchor, constant: 8),
-            label.trailingAnchor.constraint(equalTo: container.trailingAnchor, constant: -indicatorPaddingH),
-            label.centerYAnchor.constraint(equalTo: container.centerYAnchor),
-
-            container.heightAnchor.constraint(equalToConstant: indicatorHeight),
-        ])
-
-        #if os(tvOS)
-        NSLayoutConstraint.activate([
-            container.leadingAnchor.constraint(equalTo: targetView.safeAreaLayoutGuide.leadingAnchor, constant: 80),
-            container.bottomAnchor.constraint(equalTo: targetView.safeAreaLayoutGuide.bottomAnchor, constant: -80),
-        ])
-        #else
-        NSLayoutConstraint.activate([
-            container.leadingAnchor.constraint(equalTo: targetView.safeAreaLayoutGuide.leadingAnchor, constant: 20),
-            container.bottomAnchor.constraint(equalTo: targetView.safeAreaLayoutGuide.bottomAnchor, constant: -80),
-        ])
-        #endif
-
-        self.sleepIndicatorContainer = container
-        self.sleepIndicatorLabel = label
-
-        UIView.animate(withDuration: 0.3) { container.alpha = 1 }
-    }
-
-    private func updateSleepIndicator(remaining: TimeInterval) {
-        guard let label = sleepIndicatorLabel else { return }
-        let seconds = max(0, Int(remaining.rounded(.up)))
-        let hh = seconds / 3600
-        let mm = (seconds % 3600) / 60
-        let ss = seconds % 60
-        let formatted: String
-        if hh > 0 {
-            formatted = String(format: "%d:%02d:%02d", hh, mm, ss)
-        } else {
-            formatted = String(format: "%d:%02d", mm, ss)
-        }
-        label.text = String(format: loc.localized("sleep.indicator"), formatted)
-    }
-
-    private func hideSleepIndicator() {
-        guard let container = sleepIndicatorContainer else { return }
-        UIView.animate(withDuration: 0.25, animations: { container.alpha = 0 }) { _ in
-            container.removeFromSuperview()
-        }
-        sleepIndicatorContainer = nil
-        sleepIndicatorLabel = nil
-    }
-
-    private func showSleepOverlay() {
-        #if os(tvOS)
-        // On tvOS the focus engine doesn't claim arbitrary UIView subviews added to
-        // AVPlayerViewController.view, so the custom blur card's UIButtons are
-        // unreachable with the remote. UIAlertController owns its own focus context
-        // and gives us free Siri-Remote navigation.
-        guard sleepOverlayContainer == nil, let vc = playerVC else { return }
-
-        let alert = UIAlertController(
-            title: loc.localized("sleep.prompt.title"),
-            message: loc.localized("sleep.prompt.subtitle"),
-            preferredStyle: .alert
-        )
-        alert.addAction(UIAlertAction(
-            title: loc.localized("sleep.prompt.keepWatching"),
-            style: .default,
-            handler: { [weak self] _ in
-                self?.sleepOverlayContainer = nil
-                self?.handleSleepKeepWatching()
-            }
-        ))
-        alert.addAction(UIAlertAction(
-            title: loc.localized("sleep.prompt.stop"),
-            style: .destructive,
-            handler: { [weak self] _ in
-                self?.sleepOverlayContainer = nil
-                self?.handleSleepStop()
-            }
-        ))
-
-        // Use the alert controller's view as a sentinel so `hideSleepOverlay` and
-        // cleanup paths know "an overlay is active" without us tracking another field.
-        sleepOverlayContainer = alert.view
-        vc.present(alert, animated: true)
-        return
-        #else
-        guard sleepOverlayContainer == nil, let vc = playerVC else { return }
-
-        let container = UIView()
-        container.translatesAutoresizingMaskIntoConstraints = false
-        container.backgroundColor = UIColor.black.withAlphaComponent(0.6)
-        container.alpha = 0
-
-        let card = UIView()
-        card.translatesAutoresizingMaskIntoConstraints = false
-        card.layer.cornerRadius = 20
-        card.clipsToBounds = true
-
-        let blur = UIVisualEffectView(effect: UIBlurEffect(style: .dark))
-        blur.translatesAutoresizingMaskIntoConstraints = false
-        blur.isUserInteractionEnabled = false
-        card.addSubview(blur)
-
-        let icon = UIImageView(image: UIImage(systemName: "moon.zzz.fill"))
-        icon.tintColor = .white
-        icon.translatesAutoresizingMaskIntoConstraints = false
-        icon.contentMode = .scaleAspectFit
-
-        let titleLabel = UILabel()
-        titleLabel.translatesAutoresizingMaskIntoConstraints = false
-        titleLabel.font = .systemFont(ofSize: overlayTitleSize, weight: .bold)
-        titleLabel.textColor = .white
-        titleLabel.textAlignment = .center
-        titleLabel.numberOfLines = 0
-        titleLabel.text = loc.localized("sleep.prompt.title")
-
-        let subtitleLabel = UILabel()
-        subtitleLabel.translatesAutoresizingMaskIntoConstraints = false
-        subtitleLabel.font = .systemFont(ofSize: overlaySubtitleSize, weight: .medium)
-        subtitleLabel.textColor = .white.withAlphaComponent(0.8)
-        subtitleLabel.textAlignment = .center
-        subtitleLabel.numberOfLines = 0
-        subtitleLabel.text = loc.localized("sleep.prompt.subtitle")
-
-        var keepConfig = UIButton.Configuration.plain()
-        var keepTitle = AttributedString(loc.localized("sleep.prompt.keepWatching"))
-        keepTitle.font = .systemFont(ofSize: overlayButtonSize, weight: .semibold)
-        keepTitle.foregroundColor = UIColor.black
-        keepConfig.attributedTitle = keepTitle
-        keepConfig.contentInsets = NSDirectionalEdgeInsets(top: 14, leading: 24, bottom: 14, trailing: 24)
-        keepConfig.background.backgroundColor = UIColor.white.withAlphaComponent(0.95)
-        keepConfig.background.cornerRadius = 12
-
-        let keepWatchingButton = UIButton(
-            configuration: keepConfig,
-            primaryAction: UIAction { [weak self] _ in self?.handleSleepKeepWatching() }
-        )
-        keepWatchingButton.translatesAutoresizingMaskIntoConstraints = false
-
-        var stopConfig = UIButton.Configuration.plain()
-        var stopTitle = AttributedString(loc.localized("sleep.prompt.stop"))
-        stopTitle.font = .systemFont(ofSize: overlayButtonSize, weight: .semibold)
-        stopTitle.foregroundColor = UIColor.white
-        stopConfig.attributedTitle = stopTitle
-        stopConfig.contentInsets = NSDirectionalEdgeInsets(top: 14, leading: 24, bottom: 14, trailing: 24)
-        stopConfig.background.backgroundColor = UIColor.white.withAlphaComponent(0.15)
-        stopConfig.background.cornerRadius = 12
-
-        let stopButton = UIButton(
-            configuration: stopConfig,
-            primaryAction: UIAction { [weak self] _ in self?.handleSleepStop() }
-        )
-        stopButton.translatesAutoresizingMaskIntoConstraints = false
-
-        card.addSubview(icon)
-        card.addSubview(titleLabel)
-        card.addSubview(subtitleLabel)
-        card.addSubview(keepWatchingButton)
-        card.addSubview(stopButton)
-        container.addSubview(card)
-
-        let targetView = vc.view!
-        targetView.addSubview(container)
-
-        NSLayoutConstraint.activate([
-            container.topAnchor.constraint(equalTo: targetView.topAnchor),
-            container.bottomAnchor.constraint(equalTo: targetView.bottomAnchor),
-            container.leadingAnchor.constraint(equalTo: targetView.leadingAnchor),
-            container.trailingAnchor.constraint(equalTo: targetView.trailingAnchor),
-
-            card.centerXAnchor.constraint(equalTo: container.centerXAnchor),
-            card.centerYAnchor.constraint(equalTo: container.centerYAnchor),
-            card.widthAnchor.constraint(equalToConstant: overlayCardWidth),
-
-            blur.topAnchor.constraint(equalTo: card.topAnchor),
-            blur.bottomAnchor.constraint(equalTo: card.bottomAnchor),
-            blur.leadingAnchor.constraint(equalTo: card.leadingAnchor),
-            blur.trailingAnchor.constraint(equalTo: card.trailingAnchor),
-
-            icon.topAnchor.constraint(equalTo: card.topAnchor, constant: 36),
-            icon.centerXAnchor.constraint(equalTo: card.centerXAnchor),
-            icon.widthAnchor.constraint(equalToConstant: overlayIconSize),
-            icon.heightAnchor.constraint(equalToConstant: overlayIconSize),
-
-            titleLabel.topAnchor.constraint(equalTo: icon.bottomAnchor, constant: 20),
-            titleLabel.leadingAnchor.constraint(equalTo: card.leadingAnchor, constant: 24),
-            titleLabel.trailingAnchor.constraint(equalTo: card.trailingAnchor, constant: -24),
-
-            subtitleLabel.topAnchor.constraint(equalTo: titleLabel.bottomAnchor, constant: 8),
-            subtitleLabel.leadingAnchor.constraint(equalTo: card.leadingAnchor, constant: 24),
-            subtitleLabel.trailingAnchor.constraint(equalTo: card.trailingAnchor, constant: -24),
-
-            keepWatchingButton.topAnchor.constraint(equalTo: subtitleLabel.bottomAnchor, constant: 28),
-            keepWatchingButton.leadingAnchor.constraint(equalTo: card.leadingAnchor, constant: 24),
-            keepWatchingButton.trailingAnchor.constraint(equalTo: card.trailingAnchor, constant: -24),
-
-            stopButton.topAnchor.constraint(equalTo: keepWatchingButton.bottomAnchor, constant: 12),
-            stopButton.leadingAnchor.constraint(equalTo: card.leadingAnchor, constant: 24),
-            stopButton.trailingAnchor.constraint(equalTo: card.trailingAnchor, constant: -24),
-            stopButton.bottomAnchor.constraint(equalTo: card.bottomAnchor, constant: -24),
-        ])
-
-        self.sleepOverlayContainer = container
-
-        UIView.animate(withDuration: 0.3) { container.alpha = 1 }
-        #endif
-    }
-
-    private func hideSleepOverlay() {
-        guard let container = sleepOverlayContainer else { return }
-        #if os(tvOS)
-        // tvOS path uses a presented UIAlertController — dismiss it via the controller.
-        if let presented = playerVC?.presentedViewController {
-            presented.dismiss(animated: true)
-        }
-        sleepOverlayContainer = nil
-        return
-        #else
-        UIView.animate(withDuration: 0.25, animations: { container.alpha = 0 }) { _ in
-            container.removeFromSuperview()
-        }
-        sleepOverlayContainer = nil
-        #endif
-    }
-
-    private func handleSleepKeepWatching() {
-        hideSleepOverlay()
-        playerVC?.player?.play()
-        // Restart timer with the user's configured default.
-        startSleepTimerIfNeeded()
-    }
-
-    private func handleSleepStop() {
-        hideSleepOverlay()
-        playerVC?.dismiss(animated: true)
-    }
-
-    private var indicatorFontSize: CGFloat {
-        #if os(tvOS)
-        26
-        #else
-        14
-        #endif
-    }
-
-    private var indicatorIconSize: CGFloat {
-        #if os(tvOS)
-        26
-        #else
-        16
-        #endif
-    }
-
-    private var indicatorHeight: CGFloat {
-        #if os(tvOS)
-        56
-        #else
-        36
-        #endif
-    }
-
-    private var indicatorPaddingH: CGFloat {
-        #if os(tvOS)
-        24
-        #else
-        14
-        #endif
-    }
-
-    private var overlayCardWidth: CGFloat {
-        #if os(tvOS)
-        640
-        #else
-        340
-        #endif
-    }
-
-    private var overlayIconSize: CGFloat {
-        #if os(tvOS)
-        72
-        #else
-        48
-        #endif
-    }
-
-    private var overlayTitleSize: CGFloat {
-        #if os(tvOS)
-        36
-        #else
-        22
-        #endif
-    }
-
-    private var overlaySubtitleSize: CGFloat {
-        #if os(tvOS)
-        22
-        #else
-        15
-        #endif
-    }
-
-    private var overlayButtonSize: CGFloat {
-        #if os(tvOS)
-        24
-        #else
-        16
-        #endif
-    }
-
     // MARK: - Helpers
 
     private func makePlayerItem(for info: PlaybackInfo) -> AVPlayerItem {
@@ -1576,36 +1016,17 @@ final class NativeVideoPresenter {
             forName: .cinemaxDidEnterBackground, object: nil, queue: .main
         ) { [weak self] _ in
             Task { @MainActor in
-                self?.reportPlaybackProgress()
+                self?.playbackReporter.reportBackgroundProgress()
             }
-        }
-    }
-
-    private func reportPlaybackProgress() {
-        guard let info = playbackInfo, let player = playerVC?.player else { return }
-        let positionTicks = Int(player.currentTime().seconds * 10_000_000)
-        let id = itemId
-        let client = apiClient
-        let uid = userId
-        Task.detached {
-            await client.reportPlaybackProgress(
-                itemId: id, userId: uid,
-                mediaSourceId: info.mediaSourceId, playSessionId: info.playSessionId,
-                positionTicks: positionTicks, isPaused: true, playMethod: info.playMethod
-            )
         }
     }
 
     private func cleanupPlayer() {
         removeTimeObserver()
-        hideSkipButton()
+        skipSegments.teardown()
         hideSkipToEndButton()
-        stopSleepTimer()
-        hideSleepIndicator()
-        hideSleepOverlay()
+        sleepTimer.teardown()
         hideFinishedSeriesOverlay()
-        segments = []
-        activeSegmentType = nil
         playerObservation?.invalidate()
         playerObservation = nil
         if let obs = itemEndObserver {
